edit problems Relates #39
