+++ package.json
@@ -7,6 +7,7 @@
     "test": "test"
   },
   "scripts": {
+    "test": "tape ./test/test_BE/doctors.js | tap-spec",
     "start": "node ./src/index.js",
     "dev": "nodemon ./src/index.js"
   },
@@ -21,7 +22,10 @@
     "eslint": "^5.5.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
-    "nodemon": "^1.18.4"
+    "nodemon": "^1.18.4",
+    "supertest": "^3.3.0",
+    "tap-spec": "^5.0.0",
+    "tape": "^4.9.1"
   },
   "repository": {
     "type": "git",
