+++ public/css/doctors.css
@@ -1,8 +1,4 @@
-*{
-    padding:0;
-    margin: 0;
-    font-family: lato;
-}
+
 main section .container{
     width: 90%;
     margin: 0 auto;
@@ -46,12 +42,11 @@ main section .container .doctors .doctor article span{
     text-align: center;
     margin-top: 20px;
 }
-main section .container .doctors .doctor article .doctor_name{
-    font-size:20px;
-}
+main section .container .doctors .doctor article .doctor_name , 
 main section .container .doctors .doctor article .doctor_age{
     font-size:20px;
 }
+
 main section .container .doctors .doctor article .about_doctor{
     font-size:15px;
     width: 80%;
