+++ public/css/header.css
@@ -1,8 +1,4 @@
-*{
-    padding:0;
-    margin: 0;
-    font-family: lato;
-}
+
 /***************** Header Style ****************/
 header{
     position: fixed;
