+++ public/css/public.css
@@ -1,0 +1,5 @@
+*{
+    padding:0;
+    margin: 0;
+    font-family: lato;
+}
