+++ src/controllers/doctors.js
@@ -1,6 +1,6 @@
 exports.get = (req, res) => {
   res.render('doctors', {
-    title: 'Doctors || Chat It Out', headerFound: true, footerFound: true, style: ['doctors', 'header', 'footer'], javascript: ['doctors']
+    title: 'Doctors || Chat It Out', headerFound: true, footerFound: true, style: ['doctors', 'header', 'footer'], javascript: ['doctors', 'hamburger']
     ,
   });
 };
