+++ src/views/partials/footer.hbs
@@ -1,7 +1,7 @@
   <div class="container">
     <div class="logo-container">
-      <span class="first">Chat it Out </span>
-      <span class="second">Gaza psychological service</span>
+      <div class="first">Chat it Out </div>
+      <div class="second">Gaza psychological service</div>
     </div>
 
     <nav>
