+++ src/views/partials/header.hbs
@@ -1,37 +1,24 @@
 <div class="container">
 
     <div class="logo">
-
-        <span class="first_word">
-
-            Chat it Out
-
-        </span>
-
-        <span class="second_word">
-
-            Gaza psychological service
-
-        </span>
-
+        <span class="first_word">Chat it Out</span>
+        <span class="second_word">Gaza psychological service</span>
     </div>
 
     <nav>
         <ul>
-            <li><a href="#"> Chat </a></li>
-            <li><a href="#"> Articles</a></li>
-            <li><a href="#"> About Us </a></li>
-            <li><a href="#"> Login</a></li>
-            <li><a href="#"> Sign Up</a></li>
+            <li><a href="#">Chat</a></li>
+            <li><a href="#">Articles</a></li>
+            <li><a href="#">About Us</a></li>
+            <li><a href="#">Login</a></li>
+            <li><a href="#">Sign Up</a></li>
         </ul>
-
     </nav>
 
     <div class="huhamburger" id="hamburger">
-
         <span></span>
         <span></span>
         <span></span>
-
     </div>
+
 </div>
